+++ database/migrations/2024_07_15_113633_create_merks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('photo_path')->nullable();
             $table->string('codename', 50)->nullable()->unique();
-            $table->smallInteger('barcode')->nullable();
+            $table->integer('barcode')->nullable();
             $table->timestamps();
         });
     }
